refactor(spec): extract ERROR_CATALOG into pure error-catalog module
Move the stable validation error catalog (codes, summaries, fixes) into
error-catalog.ts with zero imports. errors.ts keeps SpecError instances,
SpecValidationError, and didYouMean, re-exporting the catalog for a stable
./errors.js public path.
